third_party/blink/web_tests/platform/generic/fast/forms/select/select-state-restore-expected.txt

Test if selected options are correctly restored even if their positions were changed.

PASS $("opt-la").selected is false
PASS $("opt-tk").selected is false
PASS $("opt-os").selected is true
PASS $("opt-ak").selected is false
PASS $("opt-wy").selected is false
We don't care which one is selected because their values are identical:
PASS !!($("opt-ca-1").selected ^ $("opt-ca-2").selected) is true
PASS successfullyParsed is true

TEST COMPLETE
